# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  overdue: 'overdue',
} as const;

export interface Task {
  id: number;
  description: string;
  scheduledDate: string;
  scheduledTime: string;
  scheduledAt: string;
  status: TaskStatus;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  completedAt: string | null;
  reminderAcknowledged?: boolean;
}

export interface TaskInput {
  /** @minLength 1 */
  description: string;
  scheduledDate: string;
  scheduledTime: string;
}

export interface TaskUpdate {
  /** @minLength 1 */
  description?: string;
  scheduledDate?: string;
  scheduledTime?: string;
}

export interface TaskSchedule {
  scheduledDate: string;
  scheduledTime: string;
}

export interface DashboardSummary {
  scheduled: number;
  completed: number;
  overdue: number;
  dueSoon: number;
  upcoming: Task[];
  recentlyCompleted: Task[];
}

export type SystemHealthStatus = typeof SystemHealthStatus[keyof typeof SystemHealthStatus];


export const SystemHealthStatus = {
  online: 'online',
  offline: 'offline',
} as const;

export type SystemHealthScheduler = typeof SystemHealthScheduler[keyof typeof SystemHealthScheduler];


export const SystemHealthScheduler = {
  running: 'running',
  stopped: 'stopped',
} as const;

export type SystemHealthDatabase = typeof SystemHealthDatabase[keyof typeof SystemHealthDatabase];


export const SystemHealthDatabase = {
  connected: 'connected',
  disconnected: 'disconnected',
} as const;

export interface SystemHealth {
  status: SystemHealthStatus;
  scheduler: SystemHealthScheduler;
  database: SystemHealthDatabase;
  lastEndpoint: string;
  lastStatusCode: number;
  lastResponseAt: string;
  /** @nullable */
  responseDurationMs?: number | null;
}

export type ListTasksParams = {
status?: ListTasksStatus;
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  all: 'all',
  scheduled: 'scheduled',
  upcoming: 'upcoming',
  completed: 'completed',
  overdue: 'overdue',
} as const;