# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';

export interface DashboardSummary {
  scheduled: number;
  completed: number;
  overdue: number;
  dueSoon: number;
  upcoming: Task[];
  recentlyCompleted: Task[];
}