# lib/api-zod/src/generated/types/systemHealth.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SystemHealthDatabase } from './systemHealthDatabase';
import type { SystemHealthScheduler } from './systemHealthScheduler';
import type { SystemHealthStatus } from './systemHealthStatus';

export interface SystemHealth {
  status: SystemHealthStatus;
  scheduler: SystemHealthScheduler;
  database: SystemHealthDatabase;
  lastEndpoint: string;
  lastStatusCode: number;
  lastResponseAt: Date;
  /** @nullable */
  responseDurationMs?: number | null;
}