# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  description: string;
  scheduledDate: string;
  scheduledTime: string;
  scheduledAt: Date;
  status: TaskStatus;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  completedAt: Date | null;
  reminderAcknowledged?: boolean;
}