# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  /** @minLength 1 */
  description: string;
  scheduledDate: string;
  scheduledTime: string;
}