# lib/api-zod/src/generated/types/taskSchedule.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskSchedule {
  scheduledDate: string;
  scheduledTime: string;
}