# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskUpdate {
  /** @minLength 1 */
  description?: string;
  scheduledDate?: string;
  scheduledTime?: string;
}